requirements.txt
PyQt5>=5.15.0
click>=8.0.0
jinja2>=3.0.0
networkx>=2.0
numpy>=1.20.0
pandas>=1.3.0
Pillow>=8.0.0
pytz>=2021.1
PyYAML>=5.4.0
requests>=2.26.0
selenium>=4.0.0
six>=1.16.0
sympy>=1.8
wheel>=0.37.0
html2text~=2024.2.26
docx2txt~=0.8
filetype~=1.2.0
argparse~=1.4.0
pypdf2~=3.0.1